A new method of subdivision
The beat enters through the leftmost inlet. Actual MS come in an additional inlet and is subdivided in a method adapted Cyprian at http://www.cyprian.com.au/blogg/index.php?post/2019/06/15/CREATING-A-METRONOME-IN-PD

This method renders a stable subdivision, but not a stable overall behavior. On the first run after setting the number of subdivisions this abstraction lags and renders fewer beats than expected.